+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 - `tiny-poppler`: fix issues with extraction not working as intended for softmask/mask/stencil
 - `npdf`: adjust the text when rendering
+- `sjpegli`: avoid using `strncpy` which may not null-terminate the string properly
 
 ## [0.1.0] (2025-12-27)